// src/app/enums/zohoCallbackStatus.enum.ts
export enum zohoCallbackStatus {
  NONE = 'none',
  LIVE = 'live',
  TRIAL = 'trial',
  DUNNING = 'dunning',
  UNPAID = 'unpaid',
  NON_RENEWING = 'non_renewing',
  CANCELLED = 'cancelled',
  CREATION_FAILED = 'creation_failed',
  CANCELLED_FROM_DUNNING = 'cancelled_from_dunning',
  EXPIRED = 'expired',
  TRIAL_EXPIRED = 'trial_expired',
  FUTURE = 'future'
}